+++ src/app/auth/auth-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {LoginComponent} from './pages/login/login/login.component';
+import {LoginComponent} from './pages/login/login.component';
 
 const routes: Routes = [
   {path: '', component: LoginComponent}
