+++ src/app/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {AuthRoutingModule} from './auth-routing.module';
-import {LoginComponent} from './pages/login/login/login.component';
+import {LoginComponent} from './pages/login/login.component';
 import {ReactiveFormsModule} from '@angular/forms';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatInputModule} from '@angular/material/input';
