+++ src/app/core/components/header/header.component.ts
@@ -8,6 +8,9 @@ import {User} from '../../../shared/models/user.model';
 import {YoutubeService} from '../../../youtube/services/youtube.service';
 import {SearchResponse} from '../../../youtube/models/search-response.model';
 import {Observable, of} from 'rxjs';
+import {Store} from '@ngrx/store';
+import * as fromRoot from '../../../redux/reducers';
+import * as cardAction from '../../../redux/actions/cards.action';
 
 @Component({
   selector: 'app-header',
@@ -24,7 +27,8 @@ export class HeaderComponent implements OnInit {
     private youtubeService: YoutubeService,
     public authUserService: AuthUserService,
     private data: DataService,
-    public userHolderService: UserHolderService
+    public userHolderService: UserHolderService,
+    private store: Store<fromRoot.State>
   ) {
     of(this.userHolderService)
       .subscribe((service) => {
@@ -48,6 +52,7 @@ export class HeaderComponent implements OnInit {
 
   public search(query: string): void {
     if (query.trim().length > 3) {
+      this.store.dispatch(new cardAction.SearchCards(query));
       const result: Observable<SearchResponse> | void = this.youtubeService.searchYouTubeData(query);
       if (result instanceof Observable) {
         result.subscribe((response: SearchResponse) => {
